requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.1.0
nest-asyncio==1.6.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
stripe==11.6.0
Telethon==1.28.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
